+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixes a regression where a 404 status code during deletion wasn't considered deleted. (https://github.com/pulumi/pulumi-docker-build/issues/849)
 - Fixed `exec: true` builds failing with `exit status 125` due to malformed buildx arguments. (https://github.com/pulumi/pulumi-docker-build/issues/656)
+- The `secrets` input is now correctly marked as a Pulumi secret in state. (https://github.com/pulumi/pulumi-docker-build/pull/919)
 
 ## 0.0.15 (2025-10-17)